Greatly Improved Critical effects
There is no need for you to check as how vanilla does those calculations when there is an event that already detects when a player lands a critical hit. This should make stuff far easier and also make it work with all mods that modify how critical hit triggers, like making it random instead of the vanilla behaviour.